ACP2E-2704: Getting Unable to send the cookie. Size of 'mage-messages' while trying to Reorder

+++ app/code/Magento/Quote/Model/Quote.php
@@ -1699,7 +1699,9 @@ class Quote extends AbstractExtensibleModel implements \Magento\Quote\Api\Data\C
 
             // collect errors instead of throwing first one
             if ($item->getHasError()) {
-                $this->deleteItem($item);
+                if (!$request->getForceAddToCart()) {
+                    $this->deleteItem($item);
+                }
                 foreach ($item->getMessage(false) as $message) {
                     if (!in_array($message, $errors)) {
                         // filter duplicate messages
